Refactor hardhat tasks
