+++ contracts/tasks/account.js
@@ -1,0 +1,100 @@
+// USDT has its own ABI because of non standard returns
+const usdtAbi = require("../test/abi/usdt.json").abi;
+const daiAbi = require("../test/abi/erc20.json");
+const tusdAbi = require("../test/abi/erc20.json");
+const usdcAbi = require("../test/abi/erc20.json");
+
+/**
+ * Prints test accounts.
+ */
+async function accounts(taskArguments, hre, privateKeys) {
+  const accounts = await hre.ethers.getSigners();
+  const roles = ["Deployer", "Governor"];
+
+  const isMainnetOrRinkeby = ["mainnet", "rinkeby"].includes(hre.network.name);
+  if (isMainnetOrRinkeby) {
+    privateKeys = [process.env.DEPLOYER_PK, process.env.GOVERNOR_PK];
+  }
+
+  let i = 0;
+  for (const account of accounts) {
+    const role = roles.length > i ? `[${roles[i]}]` : "";
+    const address = await account.getAddress();
+    console.log(address, privateKeys[i], role);
+    if (!address) {
+      throw new Error(`No address defined for role ${role}`);
+    }
+    i++;
+  }
+}
+
+/**
+ * Funds test accounts.
+ * @returns {Promise<void>}
+ */
+async function fund(taskArguments, hre) {
+  const addresses = require("./utils/addresses");
+  const {
+    usdtUnits,
+    daiUnits,
+    usdcUnits,
+    tusdUnits,
+    isFork,
+  } = require("./test/helpers");
+
+  let usdt, dai, tusd, usdc, nonStandardToken;
+  if (isFork) {
+    usdt = await hre.ethers.getContractAt(usdtAbi, addresses.mainnet.USDT);
+    dai = await hre.ethers.getContractAt(daiAbi, addresses.mainnet.DAI);
+    tusd = await hre.ethers.getContractAt(tusdAbi, addresses.mainnet.TUSD);
+    usdc = await hre.ethers.getContractAt(usdcAbi, addresses.mainnet.USDC);
+  } else {
+    usdt = await hre.ethers.getContract("MockUSDT");
+    dai = await hre.ethers.getContract("MockDAI");
+    tusd = await hre.ethers.getContract("MockTUSD");
+    usdc = await hre.ethers.getContract("MockUSDC");
+    nonStandardToken = await hre.ethers.getContract("MockNonStandardToken");
+  }
+
+  let binanceSigner;
+  const signers = await hre.ethers.getSigners();
+
+  if (isFork) {
+    await hre.network.provider.request({
+      method: "hardhat_impersonateAccount",
+      params: [addresses.mainnet.Binance],
+    });
+    binanceSigner = await hre.ethers.provider.getSigner(
+      addresses.mainnet.Binance
+    );
+  }
+
+  for (let i = 0; i < 10; i++) {
+    console.log(`Funding account ${i}`);
+    if (isFork) {
+      await dai
+        .connect(binanceSigner)
+        .transfer(await signers[i].getAddress(), daiUnits("1000"));
+      await usdc
+        .connect(binanceSigner)
+        .transfer(await signers[i].getAddress(), usdcUnits("1000"));
+      await usdt
+        .connect(binanceSigner)
+        .transfer(await signers[i].getAddress(), usdtUnits("1000"));
+      await tusd
+        .connect(binanceSigner)
+        .transfer(await signers[i].getAddress(), tusdUnits("1000"));
+    } else {
+      await dai.connect(signers[i]).mint(daiUnits("1000"));
+      await usdc.connect(signers[i]).mint(usdcUnits("1000"));
+      await usdt.connect(signers[i]).mint(usdtUnits("1000"));
+      await tusd.connect(signers[i]).mint(tusdUnits("1000"));
+      await nonStandardToken.connect(signers[i]).mint(usdtUnits("1000"));
+    }
+  }
+}
+
+module.exports = {
+  accounts,
+  fund,
+}
